+++ docs/ReleaseNotes.rst
@@ -31,6 +31,9 @@ infrastructure are described first, followed by tool-specific sections.
 Improvements to clang-tidy
 --------------------------
 
+- The 'nodecpp-new-array-expr' check was renamed to `nodecpp-naked-ptr-from-return
+  <http://clang.llvm.org/extra/clang-tidy/checks/nodecpp-naked-ptr-from-return.html>`_
+
 - New `nodecpp-naked-ptr-assignment
   <http://clang.llvm.org/extra/clang-tidy/checks/nodecpp-naked-ptr-assignment.html>`_ check